Created the mian config file, providers tf and the output that will give me the password for the windows instance and the ips that i can use to RDP and SSH into.
Curretnly figuring out the best way to cnsole in to the windows instance from my computer without installing a bunch of stuff.

I had an issue earlier and that was using the entry "security_groups" instead of vpc_security_group_ids. I was able to fix that and now the instance is being created.
But it is important that if the instance belongs to a vpc then we need to use vpc_security_group_ids

Next after i am able to cionsole in tommorow, i will configure binf on th edns server and server dns request from the local windows machine.

I will be using the offical redhat documentaion and some other additional docs.